Fix some sign compare warnings
git-svn-id: https://svn.apache.org/repos/asf/lucy/trunk@1328871 13f79535-47bb-0310-9956-ffa450edef68

+++ core/Lucy/Analysis/Normalizer.c
@@ -95,7 +95,7 @@ Normalizer_transform(Normalizer *self, Inversion *inversion) {
         len = utf8proc_reencode(buffer, len, self->options);
 
         if (len >= 0) {
-            if (len > token->len) {
+            if (len > (ssize_t)token->len) {
                 FREEMEM(token->text);
                 token->text = (char*)MALLOCATE(len + 1);
             }
